refactor: move apiUtils file in services folder
